Removed the trait, which is convenient but not backward compatible with php5.*
